Made the repository usable
My last attempts at making this codebase usable was nothing short of
dissappointing, but now, I've certainly hit the spot, sought of. Using
click, moving files to the expected locations, a configuration
representation and a few utilities to help with the repetitive tasks.
That is all folks, see you later.