Read on CSS grids in chapter 2 tinycss projects
Finally, a good explanation of what CSS grids are, and what `fr` unit
is. With an added benefit of two CSS functions repeat and minmax. The
grid module in css has a bunch of rules (around 8). I will experiment on
them later this week on index2.html as I did in chapter 1.