Finished chapter 6, profile card.
Went an extra mile and made the project responsive to smaller screens
and used `rem` everywhere, I'm don't really know if this is an advantage
but it does scale very well instead of mixing the units all over the
place. Also changed the theme from pink- to blue-ish which is really
amazing; the ease.